fixed page routes
