+++ .github/workflows/deploy.yml
@@ -5,7 +5,7 @@ on:
   # Runs on pushes targeting the default branch
   push:
     branches:
-        - main
+        - layoutComponent
         - 'task/**'
 
   # Allows you to run this workflow manually from the Actions tab
