+++ src/components/NotFound/NotFound.module.css
@@ -1,0 +1,8 @@
+.notfound__page {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+  gap: 20px;
+  height: 100vh;
+  margin-top: 50px;
+}
