+++ src/components/header/Header.component.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
 
 import headerLogin from '@/assets/icons/login-icon.svg';
 import headerLogo from '@/assets/icons/main_logo-icon.svg';
@@ -8,22 +9,19 @@ import headerThemeElement from '@/assets/icons/theme_mode_element-icon.svg';
 import styles from './Header.module.css';
 
 interface HeaderProps {
-    onShowAbout: () => void;
-    onShowProducts: () => void;
     showTheme: () => void;
     theme: 'light' | 'dark';
-    isShowPage: 'About' | 'ProductsList';
 }
 
-export default function Header({ onShowAbout, onShowProducts, showTheme, theme, isShowPage }: HeaderProps) {
+export default function Header({ showTheme, theme }: HeaderProps) {
     return (
         <header className={styles.header}>
             <div className={styles.header__container}>
                 <div className={styles.header__inner}>
                     <div className={styles.header__logo_wrapper}>
-                        <a className={styles.header__logo_link} href="/">
+                        <Link className={styles.header__logo_link} to="/">
                             <img className={styles.header__logo} src={headerLogo} width="40px" height="40px" alt="Logo" />
-                        </a>
+                        </Link>
                         <div className={styles.header__theme}>
                             <button
                                 className={`${styles.header__theme_button_light} ${theme === 'light' ? styles.activeTheme : ''}`}
@@ -60,20 +58,14 @@ export default function Header({ onShowAbout, onShowProducts, showTheme, theme, 
                     <nav className={styles.header__navigation}>
                         <ul className={styles.navigation__list}>
                             <li className={styles.navigation__item}>
-                                <button
-                                    onClick={onShowAbout}
-                                    className={`${styles.navigation__link} ${isShowPage === 'About' ? styles.activePage : ''}`}
-                                >
+                                <Link className={styles.navigation__link} to="/">
                                     About
-                                </button>
+                                </Link>
                             </li>
                             <li className={styles.navigation__item}>
-                                <button
-                                    onClick={onShowProducts}
-                                    className={`${styles.navigation__link} ${isShowPage === 'ProductsList' ? styles.activePage : ''}`}
-                                >
+                                <Link className={styles.navigation__link} to="products">
                                     Products
-                                </button>
+                                </Link>
                             </li>
                         </ul>
